Moved Configuration methods to TMDbClientConfiguration
