+++ TMDbLib/Client/TMDbClientConfiguration.cs
@@ -2,6 +2,8 @@
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using TMDbLib.Objects.Countries;
+using TMDbLib.Objects.Languages;
 using TMDbLib.Objects.Timezones;
 using TMDbLib.Rest;
 
@@ -9,6 +11,24 @@ namespace TMDbLib.Client
 {
     public partial class TMDbClient
     {
+        public async Task<List<Country>> GetCountriesAsync(CancellationToken cancellationToken = default(CancellationToken))
+        {
+            RestRequest req = _client.Create("configuration/countries");
+
+            RestResponse<List<Country>> response = await req.ExecuteGet<List<Country>>(cancellationToken).ConfigureAwait(false);
+
+            return (await response.GetDataObject().ConfigureAwait(false));
+        }
+
+        public async Task<List<Language>> GetLanguagesAsync(CancellationToken cancellationToken = default(CancellationToken))
+        {
+            RestRequest req = _client.Create("configuration/languages");
+
+            RestResponse<List<Language>> response = await req.ExecuteGet<List<Language>>(cancellationToken).ConfigureAwait(false);
+
+            return (await response.GetDataObject().ConfigureAwait(false));
+        }
+
         /// <summary>
         /// FindAsync movies, people and tv shows by an external id.
         /// The following trypes can be found based on the specified external id's
